+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Version 1.7.4 (in development)
 
 - Fixed class name collision upon add-on registration. This is checked since Blender 2.79.
+- Fix [T48852](https://developer.blender.org/T48852): Screenshot no longer shows "Communicating with
+  Blender Cloud".
 
 
 ## Version 1.7.3 (2017-08-08)